requirements.txt
# Core Flask Application
Flask==3.1.1
flask-cors==6.0.0
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9

# Authentication & Security
Flask-JWT-Extended==4.6.0
bcrypt==4.2.1

# Email Service
Flask-Mail==0.10.0

# Environment Configuration
python-dotenv==1.0.1

# PDF Processing
pdfminer.six==20250506
Unidecode==1.4.0

# Machine Learning (NER Model)
torch==2.7.0
transformers==4.52.3

# HTTP Requests
requests==2.32.3

# Configuration
PyYAML==6.0.2
